+++ R/RLearner.R
@@ -46,10 +46,10 @@
 #'   Default is \code{character(0)}.
 #' @return [\code{\link{RLearner}}]. The specific subclass is one of \code{\link{RLearnerClassif}},
 #'   \code{\link{RLearnerCluster}}, \code{\link{RLearnerMultilabel}},
-#'   \code{\link{RLearnerRegr}}, \code{\link{RLearnerSurv}}.
+#'   \code{\link{RLearnerRegr}}, \code{\link{RLearnerSurv}}, \code{\link{RLearnerOneClass}}.
 #' @name RLearner
 #' @rdname RLearner
-#' @aliases RLearnerClassif RLearnerCluster RLearnerMultilabel RLearnerRegr RLearnerSurv
+#' @aliases RLearnerClassif RLearnerCluster RLearnerMultilabel RLearnerRegr RLearnerSurv RLearnerOneClass
 NULL
 
 #' @export
@@ -67,7 +67,7 @@ makeRLearnerInternal = function(id, type, package, par.set, par.vals, properties
   requirePackages(package, why = stri_paste("learner", id, sep = " "), default.method = "load")
 
   assertString(id)
-  assertChoice(type, choices = c("classif", "regr", "multilabel", "surv", "cluster", "costsens"))
+  assertChoice(type, choices = c("oneclass", "classif", "regr", "multilabel", "surv", "cluster", "costsens"))
   assertSubset(properties, listLearnerProperties(type))
   assertClass(par.set, classes = "ParamSet")
   checkListElementClass(par.set$pars, "LearnerParam")
@@ -164,4 +164,13 @@ makeRLearnerCostSens = function(cl, package, par.set, par.vals = list(), propert
   )
 
   return(lrn)
+}
+
+#' @export
+#' @rdname RLearner
+makeRLearnerOneClass = function(cl, package, par.set, par.vals = list(), properties = character(0L), name = cl, short.name = cl, note = "", callees = character(0L)) {
+  addClasses(
+    makeRLearnerInternal(cl, "oneclass", package, par.set, par.vals, properties, name, short.name, note, callees),
+    c(cl, "RLearnerOneClass")
+  )
 }