+++ R/ResampleInstance.R
@@ -84,6 +84,7 @@ makeResampleInstance = function(desc, task, size, coords,  ...) {
     if (desc$stratify) {
       td = getTaskDesc(task)
       stratify.cols = switch(td$type,
+        "oneclass" = getTaskTargetNames(task),
         "classif" = getTaskTargetNames(task),
         "surv" = getTaskTargetNames(task)[2L],
         stopf("Stratification for tasks of type '%s' not supported", td$type))