+++ R/Task.R
@@ -1,5 +1,5 @@
-#' @title Create a classification, regression, survival, cluster, cost-sensitive classification or
-#' multilabel task.
+#' @title Create a classification, regression, survival, cluster, cost-sensitive classification,
+#' multilabel task or one-class classification.
 #'
 #' @description
 #' The task encapsulates the data and specifies - through its subclasses -
@@ -22,9 +22,17 @@
 #' }
 #'
 #' @details
+#' Note on multilabel:
 #' For multilabel classification we assume that the presence of labels is encoded via logical
 #' columns in \code{data}. The name of the column specifies the name of the label. \code{target}
 #' is then a char vector that points to these columns.
+#'
+#' Note on one-class classification:
+#' One-class classification problem is an unsupervised learning problem, but we still require
+#' to define a target column, in order to allow supervised evalaution if labels are available.
+#' This class columns should be a factor, where the levels are the strings denoted by
+#' \code{positive} and \code{negative}, where the former denotes the name of the anomaly class
+#' and the latter the name of the normal class.
 #'
 #' If \code{spatial = TRUE} and 'SpCV' or 'SpRepCV' are selected as
 #' resampling method, variables named \code{x} and \code{y} will be used for spatial
@@ -65,8 +73,12 @@
 #'   during a resampling iteration.
 #'   Default is \code{NULL} which means no blocking.
 #' @param positive [\code{character(1)}]\cr
-#'   Positive class for binary classification (otherwise ignored and set to NA).
-#'   Default is the first factor level of the target attribute.
+#'   Positive class for binary and one-class classification (otherwise ignored and set to NA).
+#'   For one-class classification this is the name of the \dQuote{anomaly} class and there is no default.
+#'   Default is the first factor level of the target attribute).
+#' @param negative [\code{character(1)}]\cr
+#'   Negative class name, currently only used in one-class classification, here it encodes the name
+#'   of the \dQuote{normal} class.
 #' @param fixup.data [\code{character(1)}]\cr
 #'   Should some basic cleaning up of data be performed?
 #'   Currently this means removing empty factor levels for the columns.
@@ -85,7 +97,7 @@
 #' @return [\code{\link{Task}}].
 #' @name Task
 #' @rdname Task
-#' @aliases ClassifTask RegrTask SurvTask CostSensTask ClusterTask MultilabelTask
+#' @aliases ClassifTask RegrTask SurvTask CostSensTask ClusterTask MultilabelTask OneClassTask
 #' @examples
 #' if (requireNamespace("mlbench")) {
 #'   library(mlbench)
@@ -99,7 +111,24 @@
 #'   makeClassifTask(id = "myIonosphere", data = Ionosphere, target = "Class",
 #'     positive = "good", blocking = blocking)
 #'   makeClusterTask(data = iris[, -5L])
-#' }
+#'}
+#'
+#' # for anomaly create example data with 5% anomalies
+#' set.seed(123)
+#' sigma = matrix(c(2, 0, 0, 5, 0, 0), 2, 2)
+#' normal = as.data.frame(MASS::mvrnorm(n = 1000, rep(0, 2), sigma))
+#' normal$Target = "Normal"
+#'
+#' anomaly = as.data.frame(matrix(sample(size = 50 * 2, x = 20:100, replace = TRUE), 50, 2))
+#' anomaly$Target = "Anomaly"
+#' data = rbind(normal, anomaly)
+#' data = na.omit(data)
+#'
+#' # create tasks, it is required to set the positive class (anomaly class) and
+#' # the negative class (normal class) as well as the name of the target column
+#' oneclass2d.task = makeOneClassTask("one-class-2d-example", data = data,
+#' target = "Target", positive = "Anomaly", negative = "Normal")
+#'
 NULL
 
 #' Exported for internal use.