+++ R/setPredictType.R
@@ -15,6 +15,7 @@
 #'   Regression: \dQuote{response} or \dQuote{se}.
 #'   Survival: \dQuote{response} (linear predictor) or \dQuote{prob}.
 #'   Clustering: \dQuote{response} or \dQuote{prob}.
+#'   One-class classification: \dQuote{response} or \dQuote{prob}.
 #'   Default is \dQuote{response}.
 #' @template ret_learner
 #' @family predict
@@ -29,6 +30,7 @@ setPredictType = function(learner, predict.type) {
 setPredictType.Learner = function(learner, predict.type) {
   # checks should be done down here i guess, because of recursive calls in wrappers
   assertChoice(predict.type, choices = switch(learner$type,
+    oneclass = c("response", "prob"),
     classif = c("response", "prob"),
     multilabel = c("response", "prob"),
     regr = c("response", "se"),