+++ tests/testthat/helper_helpers.R
@@ -48,18 +48,15 @@ testSimple = function(t.name, df, target, train.inds, old.predicts, parset = lis
   test = df[-inds, ]
 
   lrn = do.call("makeLearner", c(list(t.name), parset))
-  # FIXME this heuristic will backfire eventually
-  if (length(target) == 0)
-    task = makeClusterTask(data = df)
-  else if (is.numeric(df[, target]))
-    task = makeRegrTask(data = df, target = target)
-  else if (is.factor(df[, target]))
-    task = makeClassifTask(data = df, target = target)
-  else if (is.data.frame(df[, target]) && is.numeric(df[, target[1L]]) && is.logical(df[, target[2L]]))
-    task = makeSurvTask(data = df, target = target)
-  else if (is.data.frame(df[, target]) && is.logical(df[, target[1L]]))
-    task = makeMultilabelTask(data = df, target = target)
-  else
+  task = switch(lrn$type,
+    cluster = makeClusterTask(data = df),
+    regr = makeRegrTask(data = df, target = target),
+    classif = makeClassifTask(data = df, target = target),
+    surv = makeSurvTask(data = df, target = target),
+    multilabel = makeMultilabelTask(data = df, target = target),
+    oneclass = makeOneClassTask(data = df, target = target, positive = oneclass.positive, negative = oneclass.negative))
+
+  if (is.null(task))
     stop("Should not happen!")
   m = try(train(lrn, task, subset = inds))
 
@@ -143,7 +140,6 @@ testProbParsets = function(t.name, df, target, train.inds, old.probs.list, parse
     testProb(t.name, df, target, train.inds, old.probs, parset)
   }
 }
-
 
 testCV = function(t.name, df, target, folds = 2, parset = list(), tune.train, tune.predict = predict) {
   requirePackages("e1071", default.method = "load")