+++ tests/testthat/helper_objects.R
@@ -1,3 +1,5 @@
+library(Matrix)
+library(MASS)
 data(Sonar, package = "mlbench", envir = environment())
 data(BreastCancer, package = "mlbench", envir = environment())
 
@@ -24,6 +26,7 @@ binaryclass.spatial.class.col = 3
 binaryclass.spatial.class.levs = levels(binaryclass.spatial.df[, binaryclass.spatial.class.col])
 binaryclass.spatial.task = makeClassifTask("binary", data = binaryclass.spatial.df, target = binaryclass.spatial.target, spatial = TRUE)
 
+data(iris, package = "datasets")
 multiclass.df = iris
 multiclass.formula = Species~.
 multiclass.target = "Species"
@@ -63,6 +66,37 @@ noclass.test.inds  = setdiff(1:150, noclass.train.inds)
 noclass.train = noclass.df[noclass.train.inds, ]
 noclass.test  = noclass.df[noclass.test.inds, ]
 noclass.task = makeClusterTask("noclass", data = noclass.df)
+
+## create data for oneclass
+# one-classification (anomaly detection)
+set.seed(123)
+sigma = matrix(c(2, 0, 0, 5, 0, 0), 2, 2)
+normal = MASS::mvrnorm(n = 100, rep(0, 2), sigma)
+colnames(normal) = paste0("V", 1:2)
+normal = as.data.frame(normal)
+normal$Target = "normal"
+
+anomaly = matrix(sample(size = 5 * 2, x = 20:100, replace = TRUE), 5, 2)
+colnames(anomaly) = paste0("V", 1:2)
+anomaly = as.data.frame(anomaly)
+anomaly$Target = "anomaly"
+data = rbind(normal, anomaly)
+data = na.omit(data)
+data$Target = factor(data$Target)
+data[, 1:2] = normalize(data[ ,1:2])
+
+oneclass.truth = data$Target
+oneclass.df = data
+oneclass.target = "Target"
+
+oneclass.col = 3
+oneclass.train.inds = c(1:100)
+oneclass.test.inds  = setdiff(seq_len(nrow(oneclass.df)), oneclass.train.inds)
+oneclass.train = oneclass.df[oneclass.train.inds, ]
+oneclass.test  = oneclass.df[oneclass.test.inds, ]
+oneclass.positive = "anomaly"
+oneclass.negative = "normal"
+oneclass.task = makeOneClassTask("oneclass", data = oneclass.df, target = oneclass.target, positive = oneclass.positive, negative = oneclass.negative)
 
 data(BostonHousing, package = "mlbench", envir = environment())
 regr.df = BostonHousing